migrated FileIO operations from C++ to Python
