+++ py/input/params.yaml
@@ -7,7 +7,7 @@ yu22: 0.007
 yu33: 0.9
 mPhi: 1000000
 muTilde: 1001000
-m1: [0.3, 2.7]
+m1: [0.3, 2.7, 25]
 m2: 1002000
 m3: 1003000
 met1: 1004000
@@ -21,7 +21,7 @@ mqt2: 1011000
 mqt3: 1012000
 mut1: 1013000
 mut2: 1014000
-mut3: [0.3, 2.7]
+mut3: [0.3, 2.7, 25]
 mdt1: 1015000
 mdt2: 1016000
 mdt3: 1017000
