--- py/plots/data-test.yaml
@@ -1,14 +1,0 @@
-mPhi: 1.05
-muTilde: 1000000.00
-cH: [4.0e-03,-8.7e+09]
-ceH_33: [2.9e-05,-1.3e+08]
-cuH_33: [-1.1e-01,1.3e+11]
-cdH_33: [4.1e-04,-1.8e+09]
-cle_3333: [-1.1e-07,7.1e+05]
-cledq_3333: [3.0e-06,-2.0e+07]
-clequ1_3333: [8.2e-04,-2.8e+09]
-cquqd1_3333: [-1.1e-02,3.9e+10]
-cqu1_3333: [-5.2e-01,4.9e+10]
-cqu8_3333: [-3.1e+00,2.9e+11]
-cqd1_3333: [-7.0e-06,4.6e+07]
-cqd8_3333: [-4.2e-05,2.8e+08]
