+++ py/utilities.py
@@ -1,166 +1,162 @@
+from operator import call
+from time import perf_counter
 import numpy as np
-import matplotlib.pyplot as plt
-import pandas as pd
-from matplotlib.colors import LogNorm
+import yaml
+import csv
 
-label_dict = {
-    'mut3': r'$m_{\tilde{t}_R}$',
-    'm1': r'$m_1$'
-    }
+#====================================================================
+# Functions for extracting a callable from a Wilson coefficient name
+#====================================================================
 
-## Functions to convert a given WC name into LaTeX
+def split_wc_name(full_name):
+    if "_" in full_name:
+        name, idxs = full_name.split("_")
+        kw = dict()
+        
+        for i in range(len(idxs)):
+            kw["i" + str(i+1)] = int(list(idxs)[i]) - 1
+        
+        return (name, kw)
 
-def texify(wc_name):
-    parts = wc_name.split("_")
-    if len(parts) == 2:
-        [name, idx] = parts
-        if name[-1] >= '1' and name[-1] <= '9':
-            idx = "{" + '(' + name[-1] + ')' + idx + "}" 
-            name = "{" + name[1:-1] + "}"
-        else:
-            name = "{" + name[1:] + "}"
-            idx = "{" + idx + "}"
-        return [name, idx]
     else:
-        name = parts[0][1:]
-        if name == "HBox":
-            name = "{" + "H\square" + "}"
-        else:
-            name = "{" + name + "}"
-        return [name]
+        return (full_name,)
 
-def texify_abs(wc_name):   
-    parts = texify(wc_name)
-    if len(parts) == 2:
-        [name, idx] = parts
-        tex = r'$|C_%s^%s\,\, ({\rm TeV}^{-2})|$'%(name,idx)
-    else:
-        tex = r'$|C_%s\,\, ({\rm TeV}^{-2})|$'%(parts[0])
+
+def eval_wc(model, wc_name, **kw):
+    wc_tuple = split_wc_name(wc_name)
+    wc = getattr(model,wc_tuple[0])
+
+    if len(wc_tuple) == 2:
+        kw.update(wc_tuple[1])
     
-    return tex
+    return call(wc,**kw)
 
-def texify_pos(wc_name):   
-    parts = texify(wc_name)
-    if len(parts) == 2:
-        [name, idx] = parts
-        tex = r'$C_%s^%s\,\, ({\rm TeV}^{-2})$'%(name,idx)
-    else:
-        tex = r'$C_%s\,\, ({\rm TeV}^{-2})$'%(parts[0])
+#==============================
+# Functions to enable File IO
+#==============================
+
+# read param information from .yaml files
+def read_param_values(filename): 
+    with open(filename,'r') as params_file:
+        param_dict = yaml.safe_load(params_file)
+        ranges_dict = dict()
+        keys_to_pop = list()
+
+    for (key, val) in param_dict.items():
+        if type(val) == list:
+            ranges_dict[key] = np.linspace(val[0],val[1],val[2])
+            keys_to_pop.append(key)
     
-    return tex
-
-def texify_neg(wc_name):   
-    parts = texify(wc_name)
-    if len(parts) == 2:
-        [name, idx] = parts
-        tex = r'$-C_%s^%s\,\, ({\rm TeV}^{-2})$'%(name,idx)
-    else:
-        tex = r'$-C_%s\,\, ({\rm TeV}^{-2})$'%(parts[0])
+    for k in keys_to_pop:
+        param_dict.pop(k)
     
-    return tex
+    return (param_dict, ranges_dict)
 
-def texify_nounits(wc_name):   
-    parts = texify(wc_name)
-    if len(parts) == 2:
-        [name, idx] = parts
-        tex = r'$|C_%s^%s|$'%(name,idx)
-    else:
-        tex = r'$|C_%s|$'%(parts[0])
+
+# read wilson coefficient names from .txt files
+def read_wc_names(filename):
+    with open(filename, 'r') as wc_file:
+        wc_names = []
+        for line in wc_file:
+            wc_names.append(line.strip())
+
+    return wc_names
+
+
+# compute all specified WCs for a fixed benchmark point
+def create_wc_dict(model,wc_names,mubarsq,order="LOOP"):
+    result_dict = dict()
+    scale = {"mubarsq": mubarsq}
     
-    return tex
-
-def tex_label(data, wc_name):
-    is_neg = [num < 0.0 for num in np.array(data[wc_name].array) ]
-    is_pos = [num >= 0.0 for num in np.array(data[wc_name].array) ]
-    if all(is_neg):
-        return texify_neg(wc_name)
-    elif all(is_pos):
-        return texify_pos(wc_name)
-    else:
-        return texify_abs(wc_name)
-    
-# function for making 2d plots
-
-def plot_fn_2d(data, param_names, wc_name, filename):
-    assert len(param_names) == 2
-
-    param_vals = []
-    for p_name in param_names:
-        param_vals.append(np.unique(data[p_name].array))
-    param_vals = np.array(param_vals)
-
-    x_axis = param_vals[0]
-    y_axis = param_vals[1]
-
-    x_label = label_dict[param_names[0]] + r'$\,\, {\rm (TeV)}$'
-    y_label = label_dict[param_names[1]] + r'$\,\, {\rm (TeV)}$'
-
-    cbar_label = tex_label(data, wc_name)
-
-    wc_vals = abs(np.array(data[wc_name].array))
-    vals = wc_vals.reshape(len(x_axis),len(y_axis))
-
-    fig, ax = plt.subplots(figsize=(8, 6))
-    img = ax.imshow(vals, cmap='afmhot', norm=LogNorm(),
-                    origin='lower', extent=[x_axis.min(), x_axis.max(), y_axis.min(), y_axis.max()])
-
-    ax.plot([x_axis.min(), x_axis.max()], [y_axis.min(), y_axis.max()], color='gray', linestyle='--', linewidth=1.1)
-    x_fill = np.linspace(x_axis.min(), x_axis.max(), 100)
-    ax.fill_between(x_fill, x_fill, y_axis.max(), color='#AAAAAA', alpha=0.4)
-
-    cbar = fig.colorbar(img, ax=ax)
-    cbar.set_label(cbar_label, fontsize=27)
-    ax.set_xlabel(x_label, fontsize=27)
-    ax.set_ylabel(y_label, fontsize=27)
-    ax.set_title(r'$g_{1} = 0.37, \quad g_{3} = 1.1$', fontsize=27, pad=16)
-    ax.set_xticks([0.5,1.0,1.5,2.0,2.5])
-    ax.set_yticks([0.5,1.0,1.5,2.0,2.5])
-    plt.tight_layout()
-    plt.savefig(filename)
-
-def collect_vals(data, param_names, bmk_pts, wc_names):
-    values = []
-    for bp in bmk_pts:
-        x, y = bp
-        d1 = data[data[param_names[0]] == x]
-        d2 = d1[d1[param_names[1]] == y]
-
-        for i, row in d2.iterrows():
-            values.append(abs(row[wc_names].values))
-    
-    return values
-
-def create_legend(param_names, bp):
-    legend = r'%s $= {\rm %s\, TeV}$, %s = ${\rm %s\, TeV}$'%(label_dict[param_names[0]], bp[0], label_dict[param_names[1]], bp[1])
-    return legend
-
-# function for making bar plots
-
-def plot_fn_bar(data, param_names, bmk_pts, wc_names, filename,**kwargs):
-    x_pos = np.arange(len(wc_names))
-    fig, ax = plt.subplots(figsize=(15, 8))
-
-    bar_width = kwargs.get('bar_width')
-    y_min = kwargs.get('y_min')
-    y_max = kwargs.get('y_max')
-
-    vals = collect_vals(data, param_names, bmk_pts, wc_names)
-
-    ax.bar(x_pos - bar_width, vals[2], width=bar_width, label=create_legend(param_names,bmk_pts[2]), color='#FFFFAA', edgecolor='black', alpha=0.85)
-    ax.bar(x_pos, vals[1], width=bar_width, label=create_legend(param_names,bmk_pts[1]), color='tab:red', edgecolor='black', alpha=0.75)
-    ax.bar(x_pos + bar_width, vals[0], width=bar_width, label=create_legend(param_names,bmk_pts[0]), color='#030764', edgecolor='black', alpha=0.85)
-
-    ax.set_yscale('log')
-    ax.set_ylim(y_min,y_max)
-    ax.set_xticks(x_pos)
-
-    ticklabels = []
     for wc in wc_names:
-        ticklabels.append(texify_nounits(wc))
+        if order=="TREE":
+            scale["hbar"] = 0
+        else:
+            scale["hbar"] = 0.006332574
+        result_dict[wc] = eval_wc(model,wc,**scale)
 
-    ax.set_xticklabels(ticklabels)  
-    ax.set_ylabel(r'${\rm TeV}^{-2}$', fontsize=24)
-    ax.legend(fontsize=20,ncol=1,framealpha=1.0,loc=4)
+    return result_dict    
 
-    plt.tight_layout()
-    plt.savefig(filename)
+
+# write WC values along with specific parameter values to a .yaml file
+def write_to_yaml(filename,model,param_dict,param_keys,wc_names,order="LOOP"):
+    output_dict = dict()
+    for k in param_keys:
+        output_dict[k] = param_dict[k]
+
+    mubarsq = param_dict["scale"]**2
+
+    if order=="SPLIT":
+        wcs_tree = create_wc_dict(model,wc_names,mubarsq,"TREE") 
+        wcs_loop = create_wc_dict(model,wc_names,mubarsq,"LOOP")
+        for k in wcs_tree:
+            output_dict[k] = {"tree": wcs_tree[k], "loop": wcs_loop[k]-wcs_tree[k]}
+    else:
+        output_dict.update(create_wc_dict(model,param_dict,wc_names,order))
+    
+    with open(filename, 'w') as out_file:
+        yaml.dump(output_dict,out_file,sort_keys=False)
+
+
+# create a grid based on combinations of parameters that vary between min, max values
+def create_combs(arrays):
+    return np.array(np.meshgrid(*arrays)).T.reshape(-1,len(arrays))
+
+# write WC values for each point on a parameter grid to a .csv file
+def write_to_csv(filename,model,param_info,wc_names,order="LOOP"):
+    mubarsq = param_info[0]["scale"]**2
+    ranges_dict = param_info[1]
+    
+    assert(order!="SPLIT")
+    assert(len(ranges_dict)!=0)
+    
+    param_keys = list(ranges_dict.keys())
+    param_combs = create_combs(list(ranges_dict.values()))
+
+    colNames = list(param_keys) + wc_names
+
+    with open(filename, 'w') as out_file:
+        writer = csv.DictWriter(out_file,fieldnames=colNames)
+        writer.writeheader()
+        
+        output_dict = dict()
+        for param_comb in param_combs:
+            for i in range(len(param_keys)):
+                output_dict[param_keys[i]] = param_comb[i]
+            
+            model.updateParams(output_dict)
+            output_dict.update(
+                create_wc_dict(model,wc_names,mubarsq,order)
+            )
+            formatted_output = {k: (f"{v:.1e}" if isinstance(v, float) else v) for k, v in output_dict.items()}
+            writer.writerow(formatted_output)
+
+#========================================================
+# Functions for benchmarking the execution time
+#========================================================
+
+# decorator for recording the value returned and the execution time of any function that returns
+def exec_time(f):
+    def wrapper(*args, **kwargs):
+        start = perf_counter()
+        result = f(*args, **kwargs)
+        stop = perf_counter()
+        t = stop - start
+        return (result, t)
+
+    return wrapper
+
+
+# decorator for printing the value and execution time of specific functions that admit a model and wc_name as arguments
+def print_exec_time(f):
+    def wrapper(model, wc_name, **kwargs):
+        val, t = f(model, wc_name, **kwargs)
+        print(f"{wc_name} = {val:.2e}, execution time = {t:.6} secs.")
+    
+    return wrapper
+
+
+@print_exec_time
+@exec_time
+def exec(model, wc_name, **kwargs):
+    return eval_wc(model, wc_name, **kwargs)
